code/requirements.txt
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
transformers==4.44.2
lightning==2.4.0 
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
matplotlib==3.9.2
datasets==2.21.0
wandb==0.17.7
ipykernel==6.29.5
ipywidgets==8.1.5
tensorboard==2.17.1
git+https://gitlab2.informatik.uni-wuerzburg.de/kallimachos/wuenlp.git
hydra-core==1.3.2
hydra_colorlog==1.2.0
sentencepiece==0.2.0
sentence-transformers==3.0.1